Refactor extractor.extractor -> extractor
